# openapi.yaml
openapi: 3.0.3
info:
  title: distrebute.com Platform API
  version: 0.1.0
  description: |
    Unified API surface for distrebute.com — the 22 UI templates wired to
    backend services. Every endpoint is owned by exactly one microservice
    (see `x-owner` extension). Gateway routes path prefixes to services.

    Auth model: short-lived JWT access tokens (15 min) + opaque refresh
    tokens via the auth-service. Passkey / WebAuthn primary, password fallback.
  contact:
    name: distrebute.com platform team
    email: api@distrebute.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local docker-compose gateway
  - url: https://api.distrebute.com/v1
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /billing/memberships/tiers/{channelId}:
    get:
      tags: [billing]
      summary: Membership tiers for a creator
      x-owner: billing-service
      x-templates: [10]
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MembershipTier' }
  /billing/memberships/subscribe:
    post:
      tags: [billing]
      summary: Subscribe to a creator membership tier
      x-owner: billing-service
      x-templates: [10]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [tierId, paymentMethodId]
              properties:
                tierId: { type: string }
                paymentMethodId: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
  /billing/super-chat:
    post:
      tags: [billing]
      summary: Send a super chat / tip during a live stream
      x-owner: billing-service
      x-templates: [11]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                streamId: { type: string }
                amount: { type: number }
                currency: { type: string }
                message: { type: string, maxLength: 200 }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuperChat' }
  /billing/gift-subs:
    post:
      tags: [billing]
      summary: Gift N memberships to random viewers in a stream
      x-owner: billing-service
      x-templates: [11]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                streamId: { type: string }
                count: { type: integer, enum: [5, 25, 50, 100] }
                tierId: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  giftedTo:
                    type: array
                    items: { type: string }
  /billing/creator/earnings:
    get:
      tags: [billing]
      summary: Creator's earnings dashboard (Studio template)
      x-owner: billing-service
      x-templates: [9]
      parameters:
        - { in: query, name: range, schema: { type: string, enum: [30d, 90d, ytd, all] } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatorEarnings' }
...
  /notifications/preferences:
    get:
      tags: [notifications]
      summary: Notification preferences (channels + frequency)
      x-owner: notification-service
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
    put:
      tags: [notifications]
      summary: Update notification preferences
      x-owner: notification-service
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationPreferences' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }

# =============================================================================
#  COMPONENTS
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MembershipTier:
      type: object
      properties:
        id: { type: string }
        channelId: { type: string }
        name: { type: string, example: "Supporter" }
        priceAmount: { type: number }
        priceCurrency: { type: string }
        billingInterval: { type: string, enum: [month, year] }
        perks:
          type: array
          items: { type: string }
        popular: { type: boolean }
    Subscription:
      type: object
      properties:
        id: { type: string }
        tierId: { type: string }
        status: { type: string, enum: [active, past_due, canceled] }
        startedAt: { type: string, format: date-time }
        renewsAt: { type: string, format: date-time }
    SuperChat:
      type: object
      properties:
        id: { type: string }
        streamId: { type: string }
        amount: { type: number }
        currency: { type: string }
        message: { type: string }
        color: { type: string, enum: [blue, pink, gold] }
        pinnedSeconds: { type: integer }
        from:
          type: object
          properties:
            handle: { type: string }
            avatarUrl: { type: string }
    CreatorEarnings:
      type: object
      properties:
        rangeLabel: { type: string }
        totals:
          type: object
          properties:
            total: { type: number }
            ads: { type: number }
            tips: { type: number }
            members: { type: number }
            merch: { type: number }
        monthlyTrend:
          type: array
          items:
            type: object
            properties:
              month: { type: string }
              total: { type: number }
        topVideos:
          type: array
          items:
            type: object
            properties:
              videoId: { type: string }
              title: { type: string }
              views: { type: integer }
              earnings: { type: number }
...
    NotificationPreferences:
      type: object
      properties:
        push:
          type: object
          properties:
            enabled: { type: boolean }
            quietHours:
              type: object
              properties:
                start: { type: string, example: "22:00" }
                end: { type: string, example: "08:00" }
        email:
          type: object
          properties:
            weeklyDigest: { type: boolean }
            creatorActivity: { type: boolean }
            securityAlerts: { type: boolean }
        smartDigest:
          type: object
          properties:
            enabled: { type: boolean }
            dailyAt: { type: string, example: "08:30" }